feat: add cache with room

+++ user/src/main/java/com/schaefer/user/data/remote/model/UserResponse.kt
@@ -3,7 +3,7 @@ package com.schaefer.user.data.remote.model
 import com.squareup.moshi.Json
 
 data class UserResponse(
-    @field:Json(name = "id") val id: Int,
+    @field:Json(name = "id") val id: Long,
     @field:Json(name = "name") val name: String,
     @field:Json(name = "img") val img: String,
     @field:Json(name = "username") val username: String
